Refactor checkCredentials Action into class
